# packages/api-client/src/gerado/modelos/conferenciaCodigoDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API do Mony
 * API do app Monitorizze e da assistente Mony.
 * OpenAPI spec version: 1
 */

export interface ConferenciaCodigoDto {
  /**
   * @maxLength 254
   * @pattern ^(?:[A-Za-z0-9_'+\-]+\.)*[A-Za-z0-9_'+\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
   */
  email: string;
  /** @pattern ^\d{6}$ */
  codigo: string;
}