# packages/api-client/src/gerado/modelos/pedidoCodigoDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API do Mony
 * API do app Monitorizze e da assistente Mony.
 * OpenAPI spec version: 1
 */

export interface PedidoCodigoDto {
  /**
   * @maxLength 254
   * @pattern ^(?:[A-Za-z0-9_'+\-]+\.)*[A-Za-z0-9_'+\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
   */
  email: string;
}